LayoutTests/inspector/debugger/setPauseOnAssertions-expected.txt

CONSOLE MESSAGE: Failing assertion
CONSOLE MESSAGE: Failing assertion
Debugger.setPauseOnAssertions


== Running test suite: Debugger.setPauseOnAssertions
-- Running test case: Debugger.setPauseOnAssertions.Disabled
PASS: Should not pause.

-- Running test case: Debugger.setPauseOnAssertions.Enabled
PASS: Should paused.
